# PERF: optimize is_scalar, is_iterator