--- a/pandas/core/dtypes/inference.py
+++ b/pandas/core/dtypes/inference.py
@@ -25,6 +25,8 @@ is_interval = lib.is_interval
 
 is_list_like = lib.is_list_like
 
+is_iterator = lib.is_iterator
+
 
 def is_number(obj) -> bool:
     """
@@ -91,40 +93,6 @@ def _iterable_not_string(obj) -> bool:
     """
 
     return isinstance(obj, abc.Iterable) and not isinstance(obj, str)
-
-
-def is_iterator(obj) -> bool:
-    """
-    Check if the object is an iterator.
-
-    For example, lists are considered iterators
-    but not strings or datetime objects.
-
-    Parameters
-    ----------
-    obj : The object to check
-
-    Returns
-    -------
-    is_iter : bool
-        Whether `obj` is an iterator.
-
-    Examples
-    --------
-    >>> is_iterator([1, 2, 3])
-    True
-    >>> is_iterator(datetime(2017, 1, 1))
-    False
-    >>> is_iterator("foo")
-    False
-    >>> is_iterator(1)
-    False
-    """
-
-    if not hasattr(obj, "__iter__"):
-        return False
-
-    return hasattr(obj, "__next__")
 
 
 def is_file_like(obj) -> bool: